requirements.txt
discord.py
pytz
openpyxl
requests
beautifulsoup4
pymysql
selenium